Extract MassPickerItem
