+++ watch/Base.lproj/Interface.storyboard
@@ -71,7 +71,7 @@
             <objects>
                 <controller identifier="add" id="xQ9-xK-vBz" customClass="AddInterfaceController" customModule="watch" customModuleProvider="target">
                     <items>
-                        <picker height="0.75" alignment="left" focusStyle="stack" indicatorMode="shownWhileFocused" id="UIA-gD-6rt">
+                        <picker height="0.75" alignment="left" indicatorMode="shownWhileFocused" id="UIA-gD-6rt">
                             <connections>
                                 <action selector="selectedMass:" destination="xQ9-xK-vBz" id="ic7-Q3-ND2"/>
                             </connections>
